requirements.txt
amqp==5.3.1
arrow==1.4.0
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
binaryornot==0.4.4
cbor2==5.7.1
celery==5.6.0
certifi==2025.6.15
cffi==2.0.0
channels==4.3.2
chardet==5.2.0
charset-normalizer==3.4.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.6.0
cron_descriptor==2.0.6
cryptography==46.0.3
daphne==4.2.1
Django==5.2.3
django-browser-reload==1.21.0
django-celery-beat==2.8.1
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-tailwind==4.2.0
django-timezone-field==7.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.3.1
gunicorn==23.0.0
h11==0.16.0
honcho==2.0.0
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
Jinja2==3.1.6
kombu==5.6.1
Markdown==3.8
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.9.0
pyOpenSSL==25.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.3
redis==7.1.0
requests==2.32.4
rich==14.2.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.4.0
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
zope.interface==8.1.1
